requirements.txt
openai
langchain
langchain-openai
langchain-community
pydantic
streamlit
python-dotenv
faiss-cpu
pypdf